requirements.txt
requests==2.28.1
python_dotenv==0.21.0
pandas==1.5.2
python-dateutil>=2.8.1
numpy>=1.24.1
pytz>=2020.1
six>=1.5